Modularized data related and render functions
